rename codemod_riverpod to riverpod_cli (#407)
