--- packages/codemod_riverpod/lib/codemod_riverpod.dart
@@ -1,3 +1,0 @@
-library codemod_riverpod;
-
-export 'src/imports.dart';
