+++ packages/riverpod_cli/bin/riverpod_cli.dart
@@ -1,0 +1,7 @@
+import 'package:riverpod_cli/riverpod_cli.dart';
+
+void main(List<String> args) {
+  final command = RiverpodCommand();
+
+  command.run(args);
+}
